Move /pageslayout folder to /functional to share fixtures
